refactor : 패키지 구조 변경
여러 도메인들이 서로 서로 통신하며 기능을 구현하기 때문에, 이를 각자의 패키지로 구분하지 않고, 하나의 패키지에 통합하여 접근 제어자를 더욱 강하게 조일 수 있도록 하였다.
