+++ src/main/java/menu/domain/Recommend.java
@@ -1,4 +1,4 @@
-package menu.domain.recommend;
+package menu.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
@@ -7,11 +7,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import menu.domain.coach.Coach;
-import menu.domain.coach.Coaches;
-import menu.domain.menu.Menu;
-import menu.domain.menu.MenuGenerator;
-import menu.domain.menu.MenuType;
 
 public class Recommend {
     private final List<MenuType> menuTypes;
