+++ src/test/java/menu/domain/MenuGeneratorTest.java
@@ -1,4 +1,4 @@
-package menu.domain.menu;
+package menu.domain;
 
 import java.util.List;
 import java.util.Map;
